+++ lib/workload/stateful/stacks/postgres-manager/compose.yml
