+++ lib/workload/stateful/stacks/postgres-manager/deploy/lambda-build.Dockerfile
