+++ lib/workload/stateful/stacks/postgres-manager/package.json
